requirements.txt
Adafruit-Blinka==8.22.4
adafruit-circuitpython-busdevice==5.2.6
adafruit-circuitpython-neopixel==6.3.10
adafruit-circuitpython-pixelbuf==2.0.3
adafruit-circuitpython-requests==2.0.2
adafruit-circuitpython-typing==1.9.4
Adafruit-PlatformDetect==3.52.3
Adafruit-PureIO==1.1.11
arandr==0.1.10
astroid==2.15.8
asttokens==2.0.4
automationhat==0.2.0
beautifulsoup4==4.9.3
blinker==1.4
blinkt==0.1.2
build==1.0.3
buttonshim==0.0.2
Cap1xxx==0.1.3
certifi==2020.6.20
chardet==4.0.0
click==7.1.2
colorama==0.4.4
colorzero==1.1
contourpy==1.1.1
cryptography==3.3.2
cupshelpers==1.0
cycler==0.11.0
dbus-python==1.2.16
dill==0.3.7
distlib==0.3.7
distro==1.5.0
docopt==0.6.2
docutils==0.16
drumhat==0.1.0
envirophat==1.0.0
ExplorerHAT==0.4.2
filelock==3.12.4
Flask==1.1.2
fonttools==4.42.1
fourletterphat==0.1.0
gpiozero==1.6.2
html5lib==1.1
idna==2.10
importlib-metadata==6.8.0
importlib-resources==6.1.0
isort==5.12.0
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.11.3
kiwisolver==1.4.5
lazy-object-proxy==1.9.0
-e git+https://github.com/jkardach/Cauldron.git@7709f05d8dfb746524f00a2d8140d5224faf187d#egg=led
logilab-common==1.8.1
lxml==4.6.3
MarkupSafe==1.1.1
matplotlib==3.8.0
mccabe==0.7.0
microdotphat==0.2.1
mote==0.0.4
motephat==0.0.3
mypy==0.812
mypy-extensions==0.4.3
numpy==1.26.0
oauthlib==3.1.0
packaging==23.1
pantilthat==0.0.7
parso==0.8.1
pexpect==4.8.0
pgzero==1.2
phatbeat==0.1.1
pianohat==0.1.0
picamera2==0.3.12
pidng==4.0.9
piexif==1.1.3
piglow==1.2.5
pigpio==1.78
Pillow==10.0.1
pipreqs==0.4.13
platformdirs==3.10.0
psutil==5.8.0
pycairo==1.16.2
pycups==2.0.1
pyftdi==0.55.0
pygame==1.9.6
Pygments==2.7.1
PyGObject==3.38.0
pyinotify==0.9.6
PyJWT==1.7.1
pylint==2.17.6
PyOpenGL==3.1.5
pyOpenSSL==20.0.1
pyparsing==3.1.1
pyproject-hooks==1.0.0
PyQt5==5.15.2
PyQt5-sip==12.8.1
pyserial==3.5
pysmbc==1.0.23
python-apt==2.2.1
python-dateutil==2.8.2
python-prctl==1.7
pyusb==1.2.1
rainbowhat==0.1.0
reportlab==3.5.59
requests==2.25.1
requests-oauthlib==1.0.0
responses==0.12.1
roman==2.0.0
RTIMULib==7.2.1
scrollphat==0.0.7
scrollphathd==1.2.1
Send2Trash==1.6.0b1
sense-hat==2.4.0
simplejpeg==1.6.4
simplejson==3.17.2
six==1.16.0
skywriter==0.0.7
sn3218==1.2.7
soupsieve==2.2.1
spidev==3.5
ssh-import-id==5.10
thonny==4.0.1
toml==0.10.1
tomli==2.0.1
tomlkit==0.12.1
touchphat==0.0.1
twython==3.8.2
typed-ast==1.4.2
typing-extensions==4.8.0
unicornhathd==0.0.4
urllib3==1.26.5
v4l2-python3==0.3.2
virtualenv==20.24.5
webencodings==0.5.1
Werkzeug==1.0.1
wrapt==1.15.0
yarg==0.1.9
zipp==3.17.0
